pyproject.toml
[project]
name = "mcp-ag2-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.10"
dependencies = [
    "mcp>=1.1.0",
    "pyautogen>=0.4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server"]